IWorkspaceModel moved in Interfaces folder.
