+++ src/AssemblySharedInfoGenerator/AssemblySharedInfo.cs
@@ -1,5 +1,4 @@
-﻿
-using System;
+﻿using System;
 using System.Reflection;
 using System.Resources;
 using System.Runtime.InteropServices;
@@ -46,7 +45,7 @@ using System.Runtime.InteropServices;
 // to distinguish one build from another. AssemblyFileVersion is specified
 // in AssemblyVersionInfo.cs so that it can be easily incremented by the
 // automated build process.
-[assembly: AssemblyVersion("0.8.2.1540")]
+[assembly: AssemblyVersion("0.8.2.1893")]
 
 
 // By default, the "Product version" shown in the file properties window is
@@ -65,4 +64,4 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyFileVersion("0.8.2.1540")]
+[assembly: AssemblyFileVersion("0.8.2.1893")]
