+++ src/DynamoCore/DynamoCore.csproj
@@ -120,7 +120,7 @@ limitations under the License.
     <Compile Include="Core\NodeGraph.cs" />
     <Compile Include="Extensions\ExtensionLoader.cs" />
     <Compile Include="Extensions\ExtensionManager.cs" />
-    <Compile Include="Models\IWorkspaceModel.cs" />
+    <Compile Include="Interfaces\IWorkspaceModel.cs" />
     <Compile Include="Rendering\DefaultRenderPackageFactory.cs" />
     <Compile Include="Extensions\ExtensionDefinition.cs" />
     <Compile Include="Extensions\IExtensionLoader.cs" />
