+++ src/DynamoPublish/Models/PublishModel.cs
@@ -1,4 +1,5 @@
 ﻿using Dynamo.Core;
+using Dynamo.Interfaces;
 using Dynamo.Models;
 using Dynamo.Publish.Configurations;
 using Dynamo.Wpf.Authentication;
